Students/01_SecondWeek/Final Version/requirements.txt
pydobotplus
pyserial
numpy
scipy
matplotlib
